+++ sources/FtpsServerAppsShared/Services/ApplicationLinks.cs
@@ -1,0 +1,9 @@
+﻿namespace FtpsServerAppsShared.Services;
+
+public class ApplicationLinks
+{
+    public const string LatestRelease = "https://github.com/drweb86/dotnet-ftps-server/releases/latest";
+    public const string ApiLatestRelease = "https://api.github.com/repos/drweb86/dotnet-ftps-server/releases/latest";
+    public const string AboutUrl = "https://github.com/drweb86/dotnet-ftps-server";
+    public const string LicenseUrl = "https://github.com/drweb86/dotnet-ftps-server/blob/main/LICENSE_APPLICATIONS.md";
+}