+++ sources/FtpsServerWindows/Commands/AboutCommand.cs
@@ -1,24 +1,22 @@
+using FtpsServerAppsShared.Services;
 using System.Diagnostics;
 using System.Windows.Input;
 
-namespace FtpsServerApp.Commands
+namespace FtpsServerApp.Commands;
+
+public class AboutCommand : ICommand
 {
-    public class AboutCommand : ICommand
+    #pragma warning disable 67
+    public event EventHandler? CanExecuteChanged;
+    #pragma warning restore 67
+
+    public bool CanExecute(object? parameter)
     {
-        private const string AboutUrl = "https://github.com/drweb86/dotnet-ftps-server";
+        return true;
+    }
 
-        #pragma warning disable 67
-        public event EventHandler? CanExecuteChanged;
-        #pragma warning restore 67
-
-        public bool CanExecute(object? parameter)
-        {
-            return true;
-        }
-
-        public void Execute(object? parameter)
-        {
-            Process.Start(new ProcessStartInfo(AboutUrl) { UseShellExecute = true });
-        }
+    public void Execute(object? parameter)
+    {
+        Process.Start(new ProcessStartInfo(ApplicationLinks.AboutUrl) { UseShellExecute = true });
     }
 }