+++ sources/FtpsServerWindows/Controls/MainMenuControl.xaml.cs
@@ -1,5 +1,4 @@
-using System;
-using System.Reflection;
+using FtpsServerAppsShared.Services;
 using System.Windows;
 using System.Windows.Controls;
 
@@ -22,8 +21,8 @@ namespace FtpsServerApp
             InitializeComponent();
 
             // Set menu item header with version
-            var version = Assembly.GetExecutingAssembly().GetName().Version;
-            AboutMenuItem.Header = $"FTPS Server - {version}";
+            var version = CopyrightInfo.Version;
+            AboutMenuItem.Header = "FTPS Server - V" + CopyrightInfo.Version.ToString(3); ;
         }
 
         private void StartStopButton_Click(object sender, RoutedEventArgs e)